requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
pymongo[srv]==3.12.1
dnspython==2.1.0
redis==4.5.1
PyJWT==2.6.0